Equation for the Biome Magnitude Delta
This is just version 1 and will be adjusted after villages are finished and balanced against the resources need to build and the rate of mining.  Changing this equation to operate off of the total resources, instead of the resource delta would be a good idea.  This information should probably go in a more visable area, but this works too.