Resources, Tests, QuickView, Tile DLL
I really should have made multiple commits of this.  Finished the Resource Calculations and put more thought into them than I should have.  Moved Test to their own project, including thoses in the Would Project.  Started a CLI layer to plug into the C# Quick Viewer.  This is intended to be both experience with CLR and a quick and light way to debug tiles and resources.  As a part of that, BaseWorld and Tiles and all that has now been wrapped into a DLL.